+++ src/main/java/techcourse/myblog/web/IndexController.java
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import techcourse.myblog.model.ArticleRepository;
+import techcourse.myblog.repository.ArticleRepository;
 
 import javax.servlet.http.HttpSession;
