+++ src/main/java/techcourse/myblog/web/UserController.java
@@ -15,7 +15,7 @@ import org.springframework.web.servlet.view.RedirectView;
 import techcourse.myblog.UserInfo;
 import techcourse.myblog.domain.User;
 import techcourse.myblog.dto.UserDto;
-import techcourse.myblog.model.UserRepository;
+import techcourse.myblog.repository.UserRepository;
 
 import javax.servlet.http.HttpSession;
 import javax.validation.groups.Default;
